+++ src/main/kotlin/com/github/sleepypanda/feesh/features/overlays/BayouTracker.kt
@@ -25,7 +25,8 @@ object BayouTracker : IResettableTracker {
 
     data class BayouTrackerData(
         val titanoboa: CatchCounterData = CatchCounterData(),
-        val titanoboaSheds: DropCounterData = DropCounterData()
+        val titanoboaSheds: DropCounterData = DropCounterData(),
+        var snakeEyes: DropCounterData = DropCounterData(),
     )
 
     const val RESET_COMMAND = "feeshResetBayouTracker"
@@ -42,6 +43,7 @@ object BayouTracker : IResettableTracker {
 
     private val titanoboa = SeaCreatures.allSeaCreatures.find { it.name == "Titanoboa" }!!
     private val titanoboaShed = RareDrops.rareDrops.find { it.itemName == "Titanoboa Shed" }!!
+    private val snakeEyes = RareDrops.rareDrops.find { it.itemName == "Snake Eyes" }!!
 
     private val gui = FeeshGui()
         .setCoordsDataKey("bayouTracker")
@@ -51,8 +53,11 @@ object BayouTracker : IResettableTracker {
             "${titanoboa.displayName}${GRAY}: ${WHITE}10 ${GRAY}catches ago ${DARK_GRAY}(${GRAY}avg: ${WHITE}500${DARK_GRAY})",
             "${GRAY}Last on: ${WHITE}1h 30m ago",
             "${titanoboaShed.displayName}s${GRAY}: ${WHITE}5",
-            "${GRAY}Last on: ${WHITE}2h 15m ago",
-            "${GRAY}Last on: ${WHITE}1 234 ${GRAY}Titanoboas ago"
+            "${GRAY}Last on: ${WHITE}20d 2h 15m ago",
+            "${GRAY}Last on: ${WHITE}223 ${GRAY}Titanoboas ago",
+            "${snakeEyes.displayName}${GRAY}: ${WHITE}10",
+            "${GRAY}Last on: ${WHITE}3h 45m ago",
+            "${GRAY}Last on: ${WHITE}10 ${GRAY}Titanoboas ago"
         ))
         .setSettingsKey { Overlays.bayouTrackerOverlay }
         .setApplyCustomStyleKey { Overlays.bayouTrackerCustomStyle }
@@ -68,13 +73,23 @@ object BayouTracker : IResettableTracker {
         EventBus.subscribe(GameClosedEvent::class, ::onGameClosed)
     }
 
+    // TODO: Remove migration code in a while
+    // This object was added later than other counters, so we need to init it manually.
+    fun initSnakeEyes() {
+        CommonUtils.runWithCatching("Failed to initialize Snake Eyes counter in Bayou tracker") {
+            PersistentDataManager.feeshData.bayouTracker.snakeEyes = DropCounterData()
+            saveData()
+        }
+    }
+
     override fun hasData(): Boolean {
-        return data.titanoboa.hasData() || data.titanoboaSheds.hasData()
+        return data.titanoboa.hasData() || data.titanoboaSheds.hasData() || data.snakeEyes.hasData()
     }
 
     override fun resetData(force: Boolean) {
         data.titanoboa.reset()
         data.titanoboaSheds.reset()
+        data.snakeEyes.reset()
         saveData(force)
     }
 
@@ -102,6 +117,7 @@ object BayouTracker : IResettableTracker {
     private fun onTitanoboa(isDoubleHook: Boolean) {
         data.titanoboa.updateAfterCatch(titanoboa.boldDisplayName)
         data.titanoboaSheds.updateAfterCatch(isDoubleHook)
+        data.snakeEyes.updateAfterCatch(isDoubleHook)
         saveData()
         updateGuiLines()
     }
@@ -110,10 +126,17 @@ object BayouTracker : IResettableTracker {
         if (!Overlays.bayouTrackerOverlay || !WorldUtils.isInSkyblock()) return
         if (WorldUtils.getWorldName() != WorldUtils.BACKWATER_BAYOU) return
 
-        if (event.itemName == titanoboaShed.itemName) {
-            data.titanoboaSheds.updateAfterDrop(titanoboaShed.boldDisplayName, titanoboa.displayName, event.magicFind)
-            saveData()
-            updateGuiLines()
+        when (event.itemName) {
+            titanoboaShed.itemName -> {
+                data.titanoboaSheds.updateAfterDrop(titanoboaShed.boldDisplayName, titanoboa.displayName, event.magicFind)
+                saveData()
+                updateGuiLines()
+            }
+            snakeEyes.itemName -> {
+                data.snakeEyes.updateAfterDrop(snakeEyes.boldDisplayName, titanoboa.displayName, event.magicFind)
+                saveData()
+                updateGuiLines()
+            }
         }
     }
 
@@ -140,6 +163,7 @@ object BayouTracker : IResettableTracker {
         lines.add(LineInfo(baseTitle))
         lines.addAll(data.titanoboa.getOverlayLines(titanoboa.displayName))
         lines.addAll(data.titanoboaSheds.getOverlayLines(titanoboaShed.displayName, titanoboa.displayName))
+        lines.addAll(data.snakeEyes.getOverlayLines(snakeEyes.displayName, titanoboa.displayName))
         gui.setLines(lines)
         gui.setButtons(listOf(getResetGuiButton { requestReset(false) }))
     }
@@ -170,6 +194,21 @@ object BayouTracker : IResettableTracker {
             data.titanoboaSheds.initDropCount(count, lastOn)
             saveData()
             ChatUtils.sendLocalChat("${GRAY}Successfully changed Titanoboa Sheds count to $count for the Bayou tracker.", true)
+        }
+    }
+
+    fun setSnakeEyes(count: Int, lastOn: Date?) {
+        CommonUtils.runWithCatching(
+            message = "Failed to set Snake Eyes.",
+            onError = {
+                ChatUtils.sendLocalChat("${RED}Failed to set Snake Eyes.", true)
+            }
+        ) {
+            if (!WorldUtils.isInSkyblock()) return
+
+            data.snakeEyes.initDropCount(count, lastOn)
+            saveData()
+            ChatUtils.sendLocalChat("${GRAY}Successfully changed Snake Eyes count to $count for the Bayou tracker.", true)
         }
     }
 }