+++ src/main/kotlin/com/github/sleepypanda/feesh/utils/data/PersistentDataManager.kt
@@ -7,6 +7,7 @@ import com.github.sleepypanda.feesh.utils.CommonUtils
 import com.github.sleepypanda.feesh.utils.FileUtils
 import com.github.sleepypanda.feesh.utils.gui.FeeshGui
 import com.github.sleepypanda.feesh.utils.enums.Alignment
+import com.github.sleepypanda.feesh.features.overlays.BayouTracker
 import com.github.sleepypanda.feesh.features.overlays.TreasureFishingTracker
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
@@ -168,6 +169,11 @@ object PersistentDataManager {
                 TreasureFishingTracker.migrateCatchesSinceLastDye()
                 FeeshMod.LOGGER.info("[Feesh] Successfully migrated catches since last Treasure Dye")
             }
+            if (jsonHasNestedKey(feeshDataFileContent, "bayouTracker") &&
+                !jsonHasNestedKey(feeshDataFileContent, "bayouTracker", "snakeEyes")) {
+                BayouTracker.initSnakeEyes()
+                FeeshMod.LOGGER.info("[Feesh] Successfully initialized Snake Eyes counter in Bayou tracker")
+            }
         }
         if (loaded != null) {
             FeeshMod.LOGGER.info("[Feesh] Successfully loaded Feesh data entries")
@@ -187,7 +193,6 @@ object PersistentDataManager {
         }
     }
 
-    
     private fun saveOverlayCoordsDataToFileAsync() {
         FileUtils.saveJsonToFileAsync(overlayCoordsFile, overlayCoordsData, gson, executor, saveLock, "Overlay coords")
     }