+++ docs/dev/CHANGELOG.md
@@ -7,6 +7,9 @@ Released on: ???
 - Added Expertise kills counter to the Items settings section, showing exact expertise counter on fishing rods [disabled by default].
 - Added option to set a keybind for bulk resetting multiple trackers' data, to quickly reset current fishing session. You can also choose what trackers to include.
 - Added option "Do not activate when wearing Trophy armor" for Fishing profit tracker [disabled by default].
+- Added **Snake Eyes** drop statistics to the Bayou tracker.
+  - To init your existing drop count: `/feeshSetTrackerDrops SNAKE_EYES <count> [YYYY-MM-DD hh:mm:ss]`
+  - Example: `/feeshSetTrackerDrops SNAKE_EYES 5 2026-06-10 23:00:00`
 
 ## Bugfixes